Can now add variables (with optional interpolation) to demo commands. Like this: commands add "cg_thirdpersonangle %0" "0" and later commands add "cg_thirdpersonangle %0" "90"
And there will be a transition between the two values.

The variable can also be started with a "$" to not interpolate between the current and last point.
Also, the variable can be a wave function as in shaders, for example

commands add "cg_thirdpersonangle %0" "sin 0 50 0 20"

Up to 10 variables are possible, numbered %0 to %9. If you want to skip a variable on a keypoint, pass an empty string like so:

commands add "cg_thirdpersonangle %1" "" "sin 0 50 0 20"

Interpolation happens between points that have the variable. If a point does not have a particular variable, it is ignored in the evaluation.