Improvements and bugfixes to the objects mechanism. Can now change most properties of nearby objects with new commands: - objects shader - objects size - objects moveHorz - objects moveVert - objects Snap - objects modulate
Shader and size allows changing the shader and size later. moveHorz and moveVert to move objects around. Snap snaps them to an angle (default 90).

Modulate is new and allows to modulate (multiply) vertex colors of the object, in essence allowing a coloring/brightening/darkening or change of transparency of objects that use a shader with rgbGen vertex and alphaGen vertex respectively.

Also now drawing little crosses in the middle of objects when not capturing. Tho they seem invisible when the object is actually visible. But it's good to see where hidden ones are that aren't currently active (due to start/end points)